WIP: play with task notifications
Not very useful as we still need to use a flag to stop both the detect
and listen tasks.

Also, the result is the same as in the previous attempt with stopping
the tasks and starting them again afterwards. Using vTaskList in the
main loop shows that the tasks are actually recreated, with a different
task number. This most likely means the audio pipeline created in
task_play messes up things so that task_listen no longer gets the
microphone input, and pausing and resuming tasks instead isn't going to
solve that.