test2: avoid calling raw_stream_read twice
Calling raw_stream_read twice on the same audio element results in 1
fragment being sent to the AFE, and the next fragment to the API. This
is the cause of the audio arriving at the API being bad.

Using audio_recorder_data_read data instead, like in the ADF WWE
example, should avoid this problem. To do so however, we need to move
the MSG_START command back to AUDIO_REC_VAD_START to avoid this error:

W (13678) AUDIO_RECORDER: Not in speeching, return 0

And unfortunately the audio arriving at the API is still bad, but at
least we now know not to call raw_stream_read twice on the same element.