test2: increase I2S output ringbuffer size
The recorder seems to have trouble keeping up with the audio input.
Since we're now sending the audio from the recorder to the API rather
than copying the audio from the I2S reader, we're sending choppy audio
to the API. Increasing the I2S output ringbuffer seems to help a lot to
avoid this.