main: run AFE on core 1
While sallow seems to work very reliably for me in its current form, it
is behaving very erratic for Kris. We suspect it might be Wi-Fi related.
As Wi-Fi is handled by core 0, this could be keeping core 0 busy and
badly influencing AFE/SR/...

Let's run it on core 1 to see if it improves anything.