main: optimize ringbuffers
We previously increased the I2S reader and SR recorder ringbuffer sizes
to avoid missing audio fragments when there is high latency to the API.
The increased ringbuffer size for the SR recorder results in the wake
word being included in what is being sent to the API.

Since we feed the data from the SR recorder to a raw stream reader,
which in turn feeds the HTTP stream writer, we can also buffer the data
in the raw stream reader ringbuffer while the HTTP stream writer is
still doing TCP/TLS handshake. Increasing the HTTP stream writer
ringbuffer did not work.

After some experimentation it seems we can reset the I2S stream
ringbuffer to the default value, and setting the SR recorder ringbuffer
to 10KB seems to be the sweet spot to not capture the wake word.