main: use ledc_set_duty_and_update
The ledc_set_duty and ledc_update_duty are not thread-safe and should
not be called from different tasks at the same time. Since we are
calling them from a callback and in app_main, it's better to use the
thread-safe function ledc_set_duty_and_update. This requires the LEDC
fade function to be installed.