container: use wildcard in git safe path
Should silence this warning:

fatal: detected dubious ownership in repository at '/opt/esp/idf/components/openthread/openthread'
To add an exception for this directory, call:

        git config --global --add safe.directory /opt/esp/idf/components/openthread/openthread

One could argue this is unsafe but since we're running this inside a
container to isolate everything from the host system, this should be an
acceptable trade-off. Explicitly adding the path of every ESP-*
submodule would be less reliable and cumbersome to maintain.