sdkconfig: enable panic on task WDT timeout
There is a memory leak somewhere. When we run out of memory, we
sometimes trigger the task WDT. Enable panic when this happens, so we
end up with a working device after reboot.

This is a partial workaround for #19 until we find and fix the leak, but
it is probably a good idea to leave it enabled. We might need to
increase CONFIG_TASK_WDT_TIMEOUT_S.