tasks: add task for printing runtime stats
This is hidden behind a config option that depends on
FREERTOS_GENERATE_RUN_TIME_STATS. This should only be enabled for
debugging purposes.

It is in its own task, because running it in the app_main loop causes a
stack overflow, so we'd have to increase the main task stack size.