slvgl: init_lvgl to new source file
As main.c is growing quite large, and we need to overhaul the display
init code to make touch work with the display being started very early
during boot, it makes sense to move some things to separate files.

Name things slvgl to avoid conflicting with the lvlg.h header.