tasks: fix timer task
When we receive an event in the timer queue it means the timer alarm
triggered after 10 seconds. Waiting 10 more seconds will result in the
display staying on longer than expected.

The whole timer logic can and should be improved further, but at least
now the display will turn off after the expected timeout.