speech_commands: add NULL check to lookup function
If for some reason our cmd_multinet char pointer array is out of sync
with the commands in the multinet partition, it's possible we try to
look up an ID that doesn't exist in the array. This can lead to
undefined behaviour. Add a NULL check to avoid this. The value returned
when we lookup an ID that doesn't exist is "INVALID".