speech_commands: add function to calculate size
We're hitting crashes when accessing the cmd_multinet array, most likely
due to not enough stack memory in the main task. Add a function to print
the total size of the cmd_multinet array, and call it during boot.

This currently triggers the crash during boot with my cmd_multinet array
being 3309 byte.