speech_commands: avoid commands > 63 characters
There is a limit ESP_MN_MAX_PHRASE_LEN in ESP-SR. Calculate if the
"TURN OFF $friendly_name" command would exceed 63 characters and drop
the entity if this is the case. Having a turn on command for an entity
buy no turn off command will be confusing, so drop the entire entity.