hass: increase WebSocket buffer size
As mentioned in the previous commit message, Home Assistent intent
responses sometimes contain a lot of data, and the responses are being
truncated. While we now better handle the situation, printing "Error!"
when Home Assistant was perfectly able to execute the intent is not
ideal. Increase the WebSocket buffer size to allow bigger responses
without truncating.

The value 4096 is still not large enough to handle the response to
"How many lights are on?" in my Home Assistant instance, but with this
value and secure WebSockets, the response is still being truncated on
4092 bytes. This does not happen when using unencrypted WebSockets, so
we could increase it further, but we would behave differently when using
HTTP vs HTTPS. Let's settle on 4096 for now, until we figure out a clean
way to handle larger JSON responses in both cases.