hass: improve WebSocket event handler
When asking Home Assistant "how many lights are on?", the intent-end
event can contain a very long JSON response, as it includes all the
light entities in the Home Assistent instance. When this happens, the
JSON response is truncated over multiple WEBSOCKET_EVENT_DATA events,
and we can no longer parse it. Since we currently handle only the
intent-end event, this results in the display not being updated, no tone
being played, and the timer to disable the display backlight not being
started.

To solve this, make the boolean a static variable, set it to false every
time we send an intent to Home Assistant over WebSocket, and do the
display update on run-end event instead. The run-end event JSON is
short and doesn't get truncated.

As we can reach run-end with a valid intent but with ok being false, the
error message "No matching HA intent" is wrong, so change it back to
"Error!".