main: trigger WAKE_END right after VAD_END
In 086dfc0fd22d ("main: trigger WAKEUP_END shortly after VAD_END"), we
reduced the time between VAD_END and WAKEUP_END from 900ms to 100ms, and
noted that setting it to 0 caused very weird behaviour. In further
testing, we learned that setting it to 1 does not result in that same
weird behaviour, so let's set it to 1 and shave of another 99ms of the
response time.