main: inform user when connecting to Wi-Fi
We're using the ESP-ADF Wi-Fi peripheral to set up the Wi-Fi connection.
This doesn't offer a way to get the number of connection attempts, or a
way to use a custom callback. Therefore just show "Connecting to Wi-Fi"
on the screen. If the ESP-Box gets stuck there it at least hints that
the problem is likely with the Wi-Fi SSID or credentials.