CMake: add env var to disable sdkconfig sanity checks
When adding the sdkconfig sanity checks, I didn't think things through.
Making config changes that fail a sanity checks results in the inability
to re-run config afterwards. There does not appear to be a CMake
variable that we can use to skip the checks when running config, so
introduce and environment variable to toggle those checks, and only
enable that variable when running `./utils.sh build`.