network: set hostname based on MAC address
This is rather ugly, as we need a valid esp_netif_t pointer to set the
hostname, but we only get this after starting the Wi-Fi peripheral.
We also need to set the hostname before the Wi-Fi is connected,
otherwise the hostname from sdkconfig will be used.

There is a chance that Wi-Fi connects before we can set the hostname,
but it's probably very small. We could check if the Wi-Fi peripheral is
connected before setting the hostname, and disconnect it, set the
hostname, and then reconnect, but this is even uglier.

We can probably improve this by not using ESP-ADF's Wi-Fi peripheral,
but that would require a lot of new code.