config: set bogus value if config_get_char key does not exist
When we call config_get_char with a key that doesn't exist in the
config, returning NULL would cause crashes due to NULL pointer
dereference, unless we validate what we get from config_get_char
everywhere we use it.

Set a bogus value if the key does not exist to avoid such crashes.