ota: add OTA support
Once you've done an OTA and use `./utils.sh flash`, you might still boot
from ota_1 partition so you will not boot what you just flashed. This
can lead to serious WTF moments. You can use otatool to force the boot
partition to ota_0.

otatool.py switch_ota_partition --slot 0

Unfortunately esptool does not support that, and otatool is part of
ESP-IDF, so this should be done in the container and we need to pass the
device node to the container again.

As we enabled app rollback in the bootloader, a full flash is required.