main: read Wi-Fi credentials from NVS
Add a state enum to track where we are in the boot phase. When we fail
to read the Wi-Fi credentials from NVS, we skip connecting to Wi-Fi and
WAS, and wait forever.

As of this commit, the NVS partition will need to have a namezone "WIFI"
with 2 keys "PSK" and "SSID", or you will no longer be able to connected
to Wi-Fi.

NVS partition can be generated with the instructions in NOTES.md, or on
https://www.linux-ipv6.be/esp-nvs/ (temp URL).

The resulting NVS partition can then be flashed with this command:
esptool.py --port /dev/ttyACM0 write_flash 0x9000 ~/tmp/nvs.bin

Alternatively, you can use https://espressif.github.io/esptool-js/.
Make sure to set the offset to 0x9000.