main: wait indefinitely after requesting config
If we're requesting a config from WAS it means we don't have a valid
config. Continuing after requesting the config doesn't make sense, as it
will cause a lot of errors. Since we're restarting after receiving the
config from WAS anyway, let's just wait indefinitely in the main task
after requesting config from WAS.