audio: check mute status during boot
When booting with mute enabled, audio_board_init failes, and throws a
whole lot of errors before it does. Check mute status before calling
audio_board_init. If mute is enabled, print a message on the display to
inform the user to unmute to continue, and check the GPIO status in a
loop until the user unmutes, then continue boot.

Fixes #192.