utils.sh: use willow branch from our ESP-ADF fork
Instead of maintaining and applying patches for ESP-ADF in the Willow
repository we will clone ESP-ADF from our fork. The willow branch has
our required changes.

We are in the process of submitting those changes upstream, but until
we've done so, and all changes are accepted, this is a much cleaner way
of maintaining and using our ESP-ADF changes.

Drop the patches directory and do_patch function since we no longer
need them.

This will require build-docker and ./utils.sh destroy.