CI: always build container
Ideally we would only build a container when there are changes to the
Dockerfile or the workflow config that builds the container.
Unfortunately GH actions is so incredibly inflexbile that this seems
impossible.

As we need to have a build container with a tag corresponding to the git
branch or tag, when we create a new branch or tag, we need to build the
container first, before we can start the dist build. But the
documentation for the push event states this:

> If you use both the branches filter and the paths filter, the workflow will only run when both filters are satisfied.

It doesn't mention anything about "OR".

So to not further lose time with GH actions, just build the container
anyway. It's a shame this is so needlessly complex that people have to
resort to workaround that waste CPU cycles.