CI: add /willow to git safe directories
Even though we run things as root inside the container and /willow is
owned by root, we get the following error:

fatal: detected dubious ownership in repository at '/willow'
To add an exception for this directory, call:

	git config --global --add safe.directory /willow

Most likely because of this, we get an empty Willow verion:

  Willow build version:

Add /willow to the safe directories to fix this.

Ideally we should not run as root, but that's something to optimize
later.