was: add EFUSE mac address to hello message
While we already have the MAC address in the hostname, this is not very
flexible. If users want to use custom hostnames, we can no longer get
the MAC address from it.

As the ESP32 does not have a unique serial number, but it is programmed
with a unique MAC address at manufacture, the MAC address from the EFUSE
is the most reliable unique identifier. If we want to keep a persistent
inventory in WAS, we will need a unique identifier.

Therefore, add the MAC address from the EFUSE in the hello message.