treewide: improve LVGL locking
We sometimes hit a deadlock when calling lvgl_port_lock(0). Calling
lvgl_port_lock with argument 0 will wait indefinitely for the semaphore
to become available, hence greatly increasing the risk to deadlock.

Limit the time to wait for the LVGL lock to 500ms to avoid deadlocking.
Since we do not wait indefinitely anymore, there is a change we will not
be able to get the lock. Hence, we should check the return status of the
lvgl_port_lock calls, and skip LVGL calls if we don't have the lock.

With this change, we will not update the screen in rare occasions, but
not locking up is far more important than correctly printing everything
to the screen.

Finally, move some non-LVGL calls outside of the conditional, to avoid
not playing an audio response or logging to the console in case we
cannot get the lock.