audio: add event queue to ESP Audio
The reason ESP Audio's async play appeared to not play anything is
because we disable the power amplifier right after the esp_audio_play
call. As this call immediately returns, we're disabling the PA before
ESP Audio actually plays the audio.

Add an event queue so we can move disabling the PA out of the play_audio
functions into an event handler.