hass: don't store hass_set_http_auth return value
Fixes the following clang-tidy warnings:
/willow/main/endpoint/hass.c:278:5: warning: Value stored to 'ret' is never read [clang-analyzer-deadcode.DeadStores]
/willow/main/endpoint/hass.c:278:5: note: Value stored to 'ret' is never read
/willow/main/endpoint/hass.c:319:5: warning: Value stored to 'ret' is never read [clang-analyzer-deadcode.DeadStores]
/willow/main/endpoint/hass.c:319:5: note: Value stored to 'ret' is never read

We print an error message in hass_set_http_auth so we don't need to
print an error message in the callers.

We might consider returning early in the caller when hass_set_http_auth
returns false, or changing the function to void if we keep the current
behaviour. Not sure which one would be best, so let's just fix the
warnings for now.