timer: force stop recorder on session timer expiry
We only stop the audio pipeline to stream to WIS when the session timer
expires. VAD_END and WAKE_END will potentially trigger afterwards. If
this happens when a new stream to WIS was already started, all kinds of
weirdness can occur. Force stop the recorder when the session timer
expires to avoid this.