CI: rework into multiple workflows
Github Actions does not support adding dependencies between multiple
workflows. For this reason, we previously defined all CI actions in a
single workflow file. As we really don't like this approach, here is
another attempt to improve the situation.

We now have a meta workflow, that will call other workflows based on
the changed files. If any file is being changed that would result in a
changed build container, we call the build_container workflow. The
build_container workflow calls the build_willow workflow once it's done.
If the changed files would not result in a changed build container, we
directly call the build_willow workflow.

This will make the CI more efficient, and faster in case we don't change
any files that would result in a changed build container. It will also
work around the limitation of Github Actions that prevent us from
publishing container images to GHCR for PRs openend from forks. This
will still be a problem when people want to submit changes to the build
container, but we're still investigating options how to fix this
properly.