Revert "treewide: disable text LVGL text scrolling"
We previously disabled LVGL text scrolling to test if this would fix the
display locking up, but we kept experiencing the lockups. We believe
commit c8561a1076be ("sdkconfig: enable LCD_RGB_ISR_IRAM_SAFE") finally
fixed the problem.

We do not want to risk this causing problems in the 0.1.0 release again,
but now that we've branched release/v0.1, let's enable text scrolling
again in main.

This reverts commit 5b46dbd7da9cff4945b6e93a6b134e709b06ce45.