treewide: use commit hash for ESP-ADF
We currently point to the willow branch of our ESP-ADF fork. As long as
we don't update that branch, this is fine, but once this branch gets
updated, rebuilding Willow would result in a different binary, even when
building a tag. This makes bisecting near impossible.

Use a commit hash instead of a branch name to avoid above problems.