sdkconfig: use sdkconfig.defaults
Use sdkconfig.defaults instead of maintaining a full sdkconfig in git.
The sdkconfig resulting from running `./utils.sh install` with this
commit should be the same as before this commit.

We can re-evaluate some of these values at a later stage, but we should
first start with identical sdkconfig. This will make bisecting easier.
If we were to change sdkconfig values in this commit, reverting the
commit would completely revert the use of sdkconfig.defaults.

Fixes: #21