treewide: add WAS wake_start and wake_end messages
The wake_start message will send the wake audio volume to WAS, needed to
implement multiwake support. When the WAS WebSocket is disconnected,
skip sending the message, as this is latency sensitive, and
re-establishing the connection would take too long. By then, the wake
session would have already expired.

Another problem with trying to reconnect on wake, is that when WAS is
still down, the connection attempt times out and slows down the entire
operation of wake, vad, stream, ... and Willow appears to hang.

Finally, Willow will try to reconnect to WAS anyway.