treewide: add ESP32-S3-Box-3 support
We now point to ESP-ADF using a tag. We must add a tag in our fork of
ESP-ADF. If we don't, and we rebase on upstream/master, the commit hash
will change, and the old hash will disappear on Github. If this happens,
our builds will fail. Using a branch for every time we bump ESP-ADF in
Willow will get ugly fast, so let's use tags for this.