audio: hide lbl_ln3 and lbl_ln4 before calling endpoint
Hiding labels after calling the endpoint could result in them being
hidden after the endpoint unhid them. This currently seems to happen for
both the openHAB and the REST endpoint.

Adding some logging confirms the label is hid in the audio code just a
few milliseconds after unhiding in the endpoint code. We don't see this
with Home Assistant as we're using a recent Home Assistant version where
the endpoint code sets up a WebSocket. The additional JSON constructing
and parsing in the Home Assistant WebSocket code causes enough delay to
not see the issue there.