CI: build from correct container image tag
When a commit doesn't trigger a container build, we build Willow in the
container tagged main. If in a previous commit in the current branch or
PR we made changes that actually require a container rebuild, CI can
fail, as one of the Espressif frameworks could be out of date.

Fix this by moving the metadata extraction out of the build-container
workflow into the meta workflow and use its outputs an inputs for the
build workflows.