WIP: M5Stack CoreS3 attempt
This is messy. The device uses an aw9523 GPIO expander, which is not
supported in ESP-IDF or ESP-ADF. The ES7210 needs to be powered on by
toggling a pin on this GPIO expander, so support for it is required.

Additionally it seems the CoreS3 requires writing to the I2C bus with
the address left shifted, otherwise I2C read/writes fail. On top of
that, the aw9523 expects 16bit I2C commands, so it requires
modifications to the ESP-ADF I2C peripheral.