was: show connection attempt on display
After doing a full flash, there is no config on the spiffs partition,
and Willow will request it from WAS. If WAS is unreachable, because it's
not running or a firewall is blocking the connection attempt, Willow
will appear to hang, with "Connecting to Wi-Fi" on the display, even if
Willow was able to connect to Wi-Fi. This is very confusing.

Print "Connecting to WAS" on the display in init_was to improve this.