treewide: request config from WAS when connected
After doing a full flash, there is no config on the spiffs partition,
and Willow will request it from WAS. If WAS is unreachable, because it's
not running or a firewall is blocking the connection attempt, this
will fail, and Willow will not try again, so even if WAS becomes
reachable afterwards, Willow will just hang and do nothing until the
user resets the device.

Improve this by moving the config request to the WAS WebSocket callback.