speech_commands: exit cleanly when WAS is unreachable
When WAS is not running or the URL in sdkconfig is incorrect, build
fails. The only reason to have the correct WAS URL in sdkconfig is to
get the Multinet commands. If the user doesn't care about Multinet, it
doesn't make much sense to require them to set the WAS URL for something
they will not use. Improve the situation by exiting the script cleanly
when the call to get the Multinet commands from WAS fails.