config: avoid confusing error message
After doing a full flash, there is no config on the spiffs partition,
and Willow will request it from WAS. We currently try to open the file,
and if it fails, this prints the following error:

E (16:10:19.221) WILLOW/CONFIG: failed to open /spiffs/user/config/willow.json

This has lead the user to wrongly conclude the missing file is a
problem on several occasions. Check if the file exists first, and print
a less confusing log message if it doesn't.