ota: drop invalid partition check
When an OTA partition is marked as bad due to it crashing, we currently
abort OTA when we're trying to flash an OTA with the same version
string. During development, we often build with the same version string
used before, while the image contains different code. If this version
string was marked invalid, we currently abort OTA and need to manually
erase OTA data or create a new git commit to be able to flash our new
image. This slows down development a bit.

As we fallback to the other OTA partition anyway, this extra check
doesn't add much value, so simply drop it.