was: drop esp_websocket_client_destroy()
We're seeing several different crashes related to the WAS WebSocket
client, like double free in esp_websocket_client_error() or while trying
to take a semaphore on the client lock. This is difficult to reproduce
and most likely timing related.

Do not manually destroy the WebSocket client, but mark the WebSocket
client to be destroyed on exit. This way, the client will be destroyed
at the end of the WebSocket task, after the connection has been cleanly
closed.

While at it, print an error message if the esp_websocket_client_close()
call failed.