audio: init amplifier and ADC separately
We currently initialize both the amplifier module and ADC with a single
call to audio_board_init(). As the mute button cuts the power to the
ADC, we must wait with this call until the user unmutes, or the
initialization will fail. Because of this, we currently can't play any
audio when mute is active, and using the WAS locate function will
actually crash Willow.

Remove this limitation by initializing the amplifier and ADC separately.