system: initialize I2C in system
To detect the touch type controller in the ESP32-S3-Box-3, we need a
handle to the I2C bus. ESP-ADF initializes this handle in
audio_board_init, but does not expose it. Therefore, initialize I2C in
system, and make the handle a global variable.

This will trigger a warning when ESP-ADF tries to initialize it again
during board init, but this is harmless.