was: allow cancel active notify_task from WAS
This is a bit tricky as in FreeRTOS we cannot get the task data from a
running task. Therefore, change notify_active to a pointer to struct
notify_data we pass to the notify task, and add a member cancel to
cancel the running task.

While at it, also free the task data memory at the end to not leak
memory.