was: drop esp_websocket_client_destroy() completely
In commit 985375c35f55 ("was: drop esp_websocket_client_destroy()")
we've dropped esp_websocket_client_destroy() in was_deinit_task(), as it
appeared to cause crashes.

We're still seeing crashes in the WebSocket client, so let's drop the
use of esp_websocket_client_destroy() completely. Instead, if the WAS
WebSocket client appears to be not connected, wait
WAS_RECONNECT_TIMEOUT_MS for it to reconnect, and if after that it's
still not connected, inform the user.