audio: add callback for HTTP stream audio element
With the existing callbacks for the HTTP streams we can only capture
HTTP response codes. If a problem occurs on a lower layer, this doesn't
trigger these callbacks. Add an extra callback for the HTTP stream audio
element, and use an enum to distinguish which audio element sent the
callback.

When an error occurs, play the error tone from SPIFFS, write an error
message on the display, and on the console.