audio: print other HTTP errors on screen
We already explicitly handle HTTP 401 and 406 from WIS, but we don't
print anything to the screen when we hit e.g. a 403 or 404. Handle these
with a generic message printed to the screen, so the user has some
indication about what is going on. Also trigger the configured audio
response.