audio: support wis_tts_url_v2 in config
To have more flexibility with the new Coqui TTS backend, we're going to
construct the TTS URL in WAS before sending it to Willow.

We will support both the old and the new way for a while, both in Willow
and WAS, so that old Willow versions will work with new WAS versions,
and new Willow version will work with old WAS versions.