ui: center align text in ui_pr_err
When mute is active during boot, we use ui_pr_err to inform the user
that unmute is required to continue. At this point, lbl_ln3 text is
center aligned, but lbl_ln4 text is left aligned.

As we want the error messages to draw attention, let's always center
align text in those 2 labels in ui_pr_err.