audio: fix session timer
In commit 349a8e47ebe7 ("treewide: define display timeout in seconds"),
we changed reset_timer to convert the timeout to microseconds inside the
function, but we forgot to change the reset_timer call for the session
timer. As a result, the session timer was waiting for millions of
seconds rather than just the number of seconds in the config.

Fixes: 349a8e47ebe7 ("treewide: define display timeout in seconds")