audio: check wake words at runtime
In preparation of OTA support for the srmodels partition we need to
remove the sdkconfig checks. Check the list of detected models at
runtime instead.

If no models are found, or the select model is not found, instruct the
user to apply settings via WAS and abort the audio initialization.