treewide: add STATE_WRITE_FLASH
Add a new STATE_WRITE_FLASH to the willow_state enum to indicate when
we're writing to flash. Ignore WAS messages when Willow is in this
state. This way, we can keep the WAS WebSocket connected during OTA,
which allows us to inform WAS if OTA succeeded or failed, or to request
the config from WAS after flashing a new srmodel partition, so that we
can update willow.json with the new wake word.