audio: increase HTTP stream timeout after send for TTS
In commit af99de418c18 ("audio: reduce HTTP Stream timeout to 2s") we
changed the HTTP Stream timeout to 2s so we can indicate early when
we're not able to connect to WIS.

Unfortunately ESP HTTP Client only has a single timeout value, and we
now also abort the HTTP connection to WIS if TTS takes longer than 2s.
We want to keep the timeout reasonably low, as having to wait for a
TTS response for too long isn't great for UX, but 2s is too short for
long TTS responses.

Leave the initial timeout at 2s, but increase it to 10s on the
HTTP_STREAM_POST_REQUEST event, which happens after the HTTP client sent
header and body to WIS, but before fetching the response.