build: make MultiNet support opt-in
We haven't supported MultiNet in any of the Willow or WAS releases.
Let's make it opt-in, rather than disabling it in CI builds. This will
solve the missing generated_cmd_multinet.h file for people trying to
build manually.

It is still possible to enable MultiNet when building manually by
setting the MULTINET environment variable, but we will not provide
support for any issues related to it. Issues will be closed without
resolution.

Closes: #382