CI: trigger workflow on push again
If we don't trigger workflow on push, we will not build and push a new
container image when a PR is merged. Because of this, the container
image with the main tag is currently based on ESP-IDF v5.1, while we
already have code in the main branch that requires ESP-IDF v5.2.

We will also not publish a release anymore when pushing a tag.

Trigger workflow on push again, but only for the main and release
branches, and for any tag.

Also add an empty newline to Dockerfile to make sure we trigger
container build and push, so that the container image with the main tag
will be based on ESP-IDF v5.2.

Fixes: 390533f1c51a ("CI: don't trigger workflow on push")