sdkconfig: enable 802.11k and 802.11v
The ESP-IDF Wi-Fi stack seems to be terrible at deciding which AP to
connect to. The Willow in my bedroom tends to connect to my outdoor AP
on boot, which has a signal strength of around -80 dBm, while there's an
AP where it has a signal strengh of around -66 dBm. Willow is
practically unusable when it's connected to the outdoor AP, due to the
low signal strength.

Enabling 802.11v allows APs that support it to steer Willow to the
better AP. Enabling 802.11k should allow the Wi-Fi stack to make better
informed decision about which AP to connect to, or when and where to
roam.

This commit enables 802.11k and 802.11v in sdkconfig. Each of them will
be enabled in the driver in a separate follow-up commit, and a torture
test will be done for each of these commits, to test possible negative
impact.

Torture test results: 1000/1000.