idf: bump LVGL to 8.4.0
This is the current most recent version of LVGL 8. In LVGL 9, the
function lv_label_set_recolor was removed in favor of lv_span.
Refactoring the code to use lv_span is going to be non-trivial. In
addition, the lv_label_set_recolor function was re-added in LVGL master,
but they did not publish a new release after that commit. Let's stick
with LVGL 8 for now.

Torture test results: 1000/1000.