CI: grant read on packages in build-willow
For some fucked up reason CI keeps failing with useless error messages
after the repo move to the new organization. The previous commit fixed
part of it, and the next failure looks suspiciously like the previous
failure, so let's try the next thing to fix it.

Seriously though, Github Actions is a total piece of fucking garbage.