idf: bump esp_lvgl_port to 2.5.0
After increasing the display buffer size to the default used in ESP-BSP,
the LVGL task no longer seemed to saturate a CPU core, but the OTA task
was no longer able to start due to not enough memory left due to the
increase of the display buffer size. This was solved by restoring
SPIRAM_MALLOC_RESERVE_INTERNAL to its default value.

Torture test results: 1000/1000.