lvgl: display buffer in SPIRAM
It appears the problem of LVGL saturating an entire CPU core isn't fixed
by the increased buffer size. When a response is long enough to trigger
scrolling of the label on the display, Willow would crash sometimes even
after the first occurence. The torture test command and response are not
long enough to trigger scrolling, so it didn't catch this.

Further increasing the buffer then resulted in the TLS component not
being able to allocate memory.

Let's try something completely different: display buffer in SPIRAM. This
didn't work when I tried initially, and the screen would draw only
partially. Then stumbled on trans_size in the esp_lvgl_port README. It
appears enabling that, makes display buffer in SPIRAM work.

Since we have plenty SPIRAM, let's just create a buffer large enough for
the entire screen, enable double buffering, and use a trans_size of 1/10
the screen size.

This is my last attempt to fix this horribly buggy scrolling feature
before I disable it.