audio: reinitialize audio on ESP_ERR_AUDIO_OPEN
When ESP Audio crashes with ESP_ERR_AUDIO_OPEN, the entire audio stack
stops working, so we no longer respond to wake word either.

Check the return value of esp_audio_play and reinitialize the entire
audio stack if it is ESP_ERR_AUDIO_OPEN.

This does not solve the actual problem of Willow not being able to play
longer TTS responses, but at least users don't need to reboot to have a
functioning Willow after this happens.

Fixing that problem is non-trivial, as ESP Audio hardcodes the RX and TX
buffers of the HTTP stream.