audio: check mute before audio init
Booting with mute active often results in the I2C bus being completely
defunct, causing log spam of I2C errors.

The mute check was previously moved after audio init in commit
20780c45c614 ("audio: initialize amp before mute check") so that WAS
locate would work even if the device is waiting to be unmuted.
Unfortunately it seems prohibitively difficult to make that work
reliably, so to not spend any more time debugging that, the move check
is moved back before audio init.

This is not ideal, but I've spent way too much time on debugging this
issue already, and I rather avoid the log spam. The mute check was moved
after audio init to solve #319, but Willow connects to WAS before the
mute check, and commands sent from WAS actually execute, there just
won't be any audio output. In case of WAS locate, the display would
still light up.

Fixes: #415