was: always send wake_start/wake_end to WAS
To not unconditionally enable WoW due to this change, we skip checking
for the wake_result message from WAS.

By always sending the wake events to WAS, we have a better understanding
of what a Willow device is doing. This is useful, and can be used by the
Home Assistant integration for audio ducking.