network: scan all Wi-Fi channels
The esp_wifi driver uses WIFI_FAST_SCAN by default, which means it will
stop scanning after it finds a matching AP. This is the main reason
Willow sometimes connect to an AP with a terrible signal strength.

Change the scan method to WIFI_ALL_CHANNEL_SCAN so that all channels
will be scanned. This seems to increase the connect time with 2 seconds,
but this is an acceptable trade-off to provide a better user experience.
Willow becomes unusable if the signal strength is too low.

Reported-by: Ciaran Jessup <ciaranj@gmail.com>