network: use WIFI_CONNECT_AP_BY_SIGNAL sort_method
A comment in the esp_wifi driver code suggests that using this
sort_method could result in connecting to an AP without encryption, if
its signal strength is higher, but now that we set the authmode
threshold to WIFI_AUTH_WPA2_PSK explicitly, we don't need to worry about
that.

Let's use WIFI_CONNECT_AP_BY_SIGNAL so that hopefully Willow always ends
up on the AP with the best signal strength.

802.11v is still useful, as e.g. when upgrading APs, depending on the
order, Willow might roam to an AP with bad signal strength, and when the
better AP comes back online, it might not roam back. 802.11v allows the
AP or controller to tell Willow to transition to the better AP.