main: set Wi-Fi failure_retry_cnt to 3
Even after changing scan_method and sort_order, a user reported Willow
would still not connect to the AP with the best signal strength. This
turns out to be due to failure on the initial connection attempt(s).

Let's set the failure_retry_cnt to 3, which for the user resulted in
Willow always ending up on the AP with the best signal strength. The
retries are sub-ms, so this shouldn't noticably increase the connect
time.

Reported-by: Ciaran Jessup <ciaranj@gmail.com>